Pizza Monitor/services/streaming/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pika==1.3.2
numpy==1.24.3
opencv-python-headless==4.8.0.76
